initial push for LOLA script examples for minial PDS4 conversion using GDAL
